refactored ingestion into a sub module
